Update to add dark mode
This code should add dark mode to the website depending on what the user has on (ios and macos system settings change the website with this css), I couldn't test it right now on your website (I have used it before and it works) just make sure it doesn't make something not work. I can help fix it.